Removing security DB call to delete engines if security is not enabled (Upload only adds engines if security enabled)
git-svn-id: svn://107.170.10.188:8080/Monolith/dev@16472 ef391c1c-ee15-45f7-b1d6-dde3d27a2169

Former-commit-id: f3d9c6934faacd145dbfa5f81f5e3d4e6845b81b

+++ src/prerna/insights/admin/DBAdminResource.java
@@ -120,15 +120,18 @@ public class DBAdminResource {
 		else if(enginesString!=null){
 			Vector<String> engines = gson.fromJson(enginesString, Vector.class);
 			UserPermissionsMasterDB permissions = new UserPermissionsMasterDB();
+			ArrayList<String> ownedEngines = permissions.getUserOwnedEngines(((User) request.getSession().getAttribute(Constants.SESSION_USER)).getId());
 			for(String engineString: engines){
 				IEngine engine = getEngine(engineString, request);
-				deleteEngine(engine, request);
 				if(this.securityEnabled) {
-					if(request.getSession().getAttribute(Constants.SESSION_USER) != null) {
+					if(ownedEngines.contains(engineString)) {
+						deleteEngine(engine, request);
 						permissions.deleteEngine(((User) request.getSession().getAttribute(Constants.SESSION_USER)).getId(), engineString);
 					} else {
-						return Response.status(400).entity("Please log in to delete databases.").build();
+						return Response.status(400).entity("You do not have access to delete this database.").build();
 					}
+				} else {
+					deleteEngine(engine, request);
 				}
 			}
 		}
